2009-11-04	Ralf Corsépius <ralf.corsepius@rtems.org>
	* ide/pcmcia_ide.c (mpc5200_pcmciaide_write_block,
	mpc5200_pcmciaide_read_block): Change "block_size" parameter to
	uint32_t to match with ide_ctrl_fns_t.

+++ c/src/lib/libbsp/powerpc/gen5200/ide/pcmcia_ide.c
@@ -446,7 +446,7 @@ void mpc5200_pcmciaide_dma_blockop(bool is_write,
 #endif /* IDE_USE_DMA */
 
 
-void mpc5200_pcmciaide_read_block(int minor, uint16_t block_size, rtems_blkdev_sg_buffer *bufs,
+void mpc5200_pcmciaide_read_block(int minor, uint32_t block_size, rtems_blkdev_sg_buffer *bufs,
                                   uint32_t *cbuf, uint32_t *pos)
 {
 
@@ -536,7 +536,7 @@ void mpc5200_pcmciaide_read_block(int minor, uint16_t block_size, rtems_blkdev_s
   }
 }
 
-void mpc5200_pcmciaide_write_block(int minor, uint16_t block_size,
+void mpc5200_pcmciaide_write_block(int minor, uint32_t block_size,
                                    rtems_blkdev_sg_buffer *bufs, uint32_t *cbuf,
                                    uint32_t *pos)
